feat(cmd): 在结构体中添加了交易时间字段

+++ internal/model/model.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `1min_trade_data` (
   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
-  `time` int(10) unsigned COMMENT '交易时间',
+  `time` bigint unsigned COMMENT '交易时间',
   `open` float(20,10) COMMENT '开盘价',
   `close` float(20,10)  COMMENT '收盘价',
   `highest_price` float(20,10)  COMMENT '最高价',
@@ -11,7 +11,7 @@ CREATE TABLE `1min_trade_data` (
 
 CREATE TABLE `5min_trade_data` (
   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
-  `time` int(10) unsigned COMMENT '交易时间',
+  `time` bigint unsigned COMMENT '交易时间',
   `open` float(20,10) COMMENT '开盘价',
   `close` float(20,10)  COMMENT '收盘价',
   `highest_price` float(20,10)  COMMENT '最高价',
